Added shitty nav bar + URL Change
Changed the position of the local repository on my PC to the Google Drive file (so when I change on my laptop, it'll auto update here). Hoping it won't mess anything up -- should be able to just go to older history though :). Git pog. Would've done more; tired af and it's 9:30PM.